RPT: r001 30
Starting this task I began looking for the countdown file that
was specified in the r001_description. I was unable to find the file in
the folder specified so I asked a classmate for the location. After
retrieving the hidden file from the bin directory I installed the
minitest gem, made countdown into a method and began writing a test for
the method.