RPT r003 30
Solved all but one code smell. Feature envy is a major problem in this
code problem because these are both utility functions without a class.
The tests should be changed to make a JustificationUtility and then
utilize it so we can make use of instance variables and avoid
referencing self more than the parameter variables. The result of
"fixing" the code smells to make reek happy in this problem actually
resulted in some very badly styled programming.