RPT: r004 30
Learned about the expect keyword and how to use it. Reverted an edit i made to the test files in my last commit and got them to work without having to edit the require statements. Learned about the .rspec config file in the root of the project. 24 failures right now using the new expect syntax, currently learning how to do comparisons with this syntax