RPT: r004 30
Learned the way to do comparisons with new expect syntax and fixed all errors in tests. 0 failures and 0 errors in the tests now.
Then used the coverage report to find the line that was not being covered inside the game.rb file and wrote a test to cover when player2 wins thegame. Achieved 100% code coverage