RPT: r005 30
In this iteration I downloaded the files and arranged them in the rspec format (spec directory). Then i had to create a spec_helper and require it in the spec test class. I also had to create a .rspec config file in the root of the project. Then in the spec_helper I enabled simple_cov and finally generated a coverage report as well as passing tests.