RPT: r007 30
Step 2 Finished. Got coverage to 97.6%, only one line remains in which I
have decided to ask for assistance because after lots of attempts at
understanding the logic in the choose_pivot method i was unable to cover
that line with many new test cases introduced. Therefore, I will move on
to step 3 until i can get assistance on that topic