RPT: r007 30
Lots of progress on Step 4. I had to go back and follow the slideshow
here https://github.com/DonSchado/colognerb-on-mutant/blob/master/mutation_testing_slides.pdf
which is one of the links on the class wiki and using this information i
was able to finally get mutant to work successfully and produce
valuable results. Now I have mutant showing 97% coverage and most
mutations being "killed". It is showing a mutation and now i will
attempt to fix it