RPT: r007 30
Continuting to improve results from mutation testing. Fixed an issue
with simplecov not being in one of my GemFiles which was causing all
mutations to have the same problem. Just ran into an issue with
mutation testing halting on the insertionSort method and never moving
forward. The mutation test reads 124/175 and never progresses. Currently
researching why this may be happening and looking for a fix.