RPT: r007 30
Using the checklist method I went back to the task of achieving 100%
coverage. I made a new test case for the pivot method and with the
refactored code (for reek) I was able to get 100% coverage. Regardless
of this the mutation test is still freezing for me when all methods are
uncommented and there does not seem to be many resources troubleshooting
this issue online. I am now going to proceed with asking another student
if they encountered this problem to not waste any more time.