RPT: r007 21
In order to eliminiate the last line that was causing alive mutations i
decided to try and optimaze the choose_pivot method by rewriting it in a
more explicit way. I did reduce my alive mutations to only 3 by doing
this but i broke one of my test cases. When the array to sort is big,
the sort is not correct. I ran out of time but next time i will fix this
test case and then have almost no alive mutations