RPT: r008 30
Next thing on checklist is convert minitest to rspec. This commit starts
and finishes that task. Had to look at rspec documentation to learn
about how to use the before block in the specification. I needed to use
it to create a new nim instance for each test. It replaces the setup
method in the minitest version of the testing file