RPT: r008 30
Started Step 3 on the checklist. Generated a coverage statistic and as
the instructions described, not much was covered by the tests. Next I
researched using the provided links about how to test I/O and began
writing an example game with I/O handling. I used the NimPlayerTesting
script that was included in the project repo to base my test off of.
Test is not complete at this time but is in good progress. Also fixed
reek errors in all code provided including the NimPlayerTesting script