RPT: r008 30
Next thing in my checklist was covering the display method and the few
remaining program flows that hadn't been traversed. In this time block i
mananged to do that and achive near full coverage. I am at 90% with all
IO tested and all logic tested. The remaining 10% is the __FILE__ method
which is only executed when the script is run directly therefore it is
not easily testable. Since that method is just running other methods
(which are tested) in order, it does not have a big impact.